requirements.txt
Pillow==9.4.0
Wand==0.6.7